+++ example/fl_mirror_example.dart
@@ -1,6 +1,0 @@
-import 'package:fl_mirror/fl_mirror.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}