+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+## [1.0.0] - 2025-06-03
 
-- Initial version.
+- Initial release with:
+- `Mirror<T>` result abstraction
+- `Failure` hierarchy
+- `Mirrorable` base class